pyproject.toml
[tool.poetry]
name = "agldt_retrain"
version = "0.1.0"
description = ""
authors = ["Caio Geraldes <caio.geraldes@usp.br>"]

[tool.poetry.dependencies]
python = "^3.10"
stanza = "^1.4.2"
GitPython = "^3.1.31"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
